fix(storage, omniread): Update repository to return pointers
This change switches the repository Find method to return a pointer to
an object rather than the object itself. This is used to easily support
the case where no database error occurred by also no object was find by
that id etc.

Additionally the superfluous WriteHeader calls were removed from the
route handlers, and the 404 not found checks were added