feat(omniread): Add post routes outline
This includes all the tests (some may need tweaks) for the post routes
CRUD operations. It also contains the implementation for reading a post.

This will need to be reworked once auth (!2) has been implemented, as we
will probably want an add post endpoint rather than a create post with
an arbitrary user id